requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
azure-core==1.34.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.9.0
openai==1.77.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
